Rename vfeature sources to HidPd.
