+++ HidBattExt/HidPd.cpp
