+++ HidBattExt/HidPd.hpp
