+++ HidBattExt/TailLight.vcxproj
@@ -133,7 +133,7 @@ copy $(OutDir)\TailLight.cer $(PackageDir)</Command>
   <ItemGroup>
     <ClCompile Include="device.cpp" />
     <ClCompile Include="driver.cpp" />
-    <ClCompile Include="vfeature.cpp" />
+    <ClCompile Include="HidPd.cpp" />
     <ResourceCompile Include="module.rc" />
   </ItemGroup>
   <ItemGroup>
@@ -145,7 +145,7 @@ copy $(OutDir)\TailLight.cer $(PackageDir)</Command>
     <ClInclude Include="device.h" />
     <ClInclude Include="driver.h" />
     <ClInclude Include="HidPdReport.h" />
-    <ClInclude Include="vfeature.h" />
+    <ClInclude Include="HidPd.hpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
 </Project>
